[#106] Moving Paper from posts to core
