+++ cms/src/core/components/Paper.js
