+++ cms/src/core/components/__tests__/Paper.spec.js
